# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 606a4e64a3e0491123244fdd
  title: SchoolManagementSystem
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /teacher:
    post:
      x-skaffolder-id: 606a526aa3e049112324509a
      x-skaffolder-name: create
      x-skaffolder-id-resource: 606a50f3a3e0491123245077
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - teacher
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/teacher"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                surname:
                  type: string
    get:
      x-skaffolder-id: 606a526aa3e04911232450a2
      x-skaffolder-name: list
      x-skaffolder-id-resource: 606a50f3a3e0491123245077
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - teacher
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/teacher"
  "/teacher/{id}":
    delete:
      x-skaffolder-id: 606a526aa3e04911232450a1
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 606a50f3a3e0491123245077
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 606a526aa3e049112324509b
      x-skaffolder-name: get
      x-skaffolder-id-resource: 606a50f3a3e0491123245077
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/teacher"
    post:
      x-skaffolder-id: 606a526aa3e0491123245099
      x-skaffolder-name: update
      x-skaffolder-id-resource: 606a50f3a3e0491123245077
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/teacher"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/teacher"
  "/teacher/findBy_courses/{key}":
    get:
      x-skaffolder-id: 606a5271a3e04911232450ac
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 606a50f3a3e0491123245077
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - teacher
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/teacher"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    teacher:
      x-skaffolder-id: 606a50f3a3e0491123245077
      x-skaffolder-id-db: 606a4e64a3e0491123244fde
      x-skaffolder-id-entity: 606a50f3a3e0491123245074
      x-skaffolder-url: /teacher
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 606a50f3a3e0491123245075
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 606a50f3a3e0491123245076
          x-skaffolder-type: String
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 606a51d7a3e0491123245087
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 606a50f3a3e0491123245074
          x-skaffolder-ent2: 606a511da3e049112324507a
      required:
        - _id
...